requirements.txt
openai==0.28
python-dotenv
fastapi
uvicorn
questionary
pytest
httpx
pytest-asyncio
coverage
pydantic
pyyaml
rich
dotenv